Move some URL extensions to Common
