+++ Tests/SharedTests/NSURLExtensionsTests.swift
@@ -70,114 +70,6 @@ class NSURLExtensionsTests: XCTestCase {
         }
     }
 
-    // MARK: Public Suffix
-    func testNormalBaseDomainWithSingleSubdomain() {
-        // TLD Entry: co.uk
-        let url = "http://a.bbc.co.uk".asURL!
-        let expected = url.publicSuffix!
-        XCTAssertEqual("co.uk", expected)
-    }
-
-    func testCanadaComputers() {
-        let url = "http://m.canadacomputers.com".asURL!
-        let actual = url.baseDomain!
-        XCTAssertEqual("canadacomputers.com", actual)
-    }
-
-    func testMultipleSuffixesInsideURL() {
-        let url = "http://com:org@m.canadacomputers.co.uk".asURL!
-        let actual = url.baseDomain!
-        XCTAssertEqual("canadacomputers.co.uk", actual)
-    }
-
-    func testNormalBaseDomainWithManySubdomains() {
-        // TLD Entry: co.uk
-        let url = "http://a.b.c.d.bbc.co.uk".asURL!
-        let expected = url.publicSuffix!
-        XCTAssertEqual("co.uk", expected)
-    }
-
-    func testWildCardDomainWithSingleSubdomain() {
-        // TLD Entry: *.kawasaki.jp
-        let url = "http://a.kawasaki.jp".asURL!
-        let expected = url.publicSuffix!
-        XCTAssertEqual("a.kawasaki.jp", expected)
-    }
-
-    func testWildCardDomainWithManySubdomains() {
-        // TLD Entry: *.kawasaki.jp
-        let url = "http://a.b.c.d.kawasaki.jp".asURL!
-        let expected = url.publicSuffix!
-        XCTAssertEqual("d.kawasaki.jp", expected)
-    }
-
-    func testExceptionDomain() {
-        // TLD Entry: !city.kawasaki.jp
-        let url = "http://city.kawasaki.jp".asURL!
-        let expected = url.publicSuffix!
-        XCTAssertEqual("kawasaki.jp", expected)
-    }
-
-    // MARK: Base Domain
-    func testNormalBaseSubdomain() {
-        // TLD Entry: co.uk
-        let url = "http://bbc.co.uk".asURL!
-        let expected = url.baseDomain!
-        XCTAssertEqual("bbc.co.uk", expected)
-    }
-
-    func testNormalBaseSubdomainWithAdditionalSubdomain() {
-        // TLD Entry: co.uk
-        let url = "http://a.bbc.co.uk".asURL!
-        let expected = url.baseDomain!
-        XCTAssertEqual("bbc.co.uk", expected)
-    }
-
-    func testBaseDomainForWildcardDomain() {
-        // TLD Entry: *.kawasaki.jp
-        let url = "http://a.b.kawasaki.jp".asURL!
-        let expected = url.baseDomain!
-        XCTAssertEqual("a.b.kawasaki.jp", expected)
-    }
-
-    func testBaseDomainForWildcardDomainWithAdditionalSubdomain() {
-        // TLD Entry: *.kawasaki.jp
-        let url = "http://a.b.c.kawasaki.jp".asURL!
-        let expected = url.baseDomain!
-        XCTAssertEqual("b.c.kawasaki.jp", expected)
-    }
-
-    func testBaseDomainForExceptionDomain() {
-        // TLD Entry: !city.kawasaki.jp
-        let url = "http://city.kawasaki.jp".asURL!
-        let expected = url.baseDomain!
-        XCTAssertEqual("city.kawasaki.jp", expected)
-    }
-
-    func testBaseDomainForExceptionDomainWithAdditionalSubdomain() {
-        // TLD Entry: !city.kawasaki.jp
-        let url = "http://a.city.kawasaki.jp".asURL!
-        let expected = url.baseDomain!
-        XCTAssertEqual("city.kawasaki.jp", expected)
-    }
-
-    func testBugzillaURLDomain() {
-        let url = "https://bugzilla.mozilla.org/enter_bug.cgi?format=guided#h=dupes|Data%20%26%20BI%20Services%20Team|"
-        let nsURL = url.asURL
-        XCTAssertNotNil(nsURL, "URL parses.")
-
-        let host = nsURL!.normalizedHost
-        XCTAssertEqual(host!, "bugzilla.mozilla.org")
-        XCTAssertEqual(nsURL!.fragment!, "h=dupes%7CData%20%26%20BI%20Services%20Team%7C")
-    }
-
-    func testIPv6Domain() {
-        let url = "http://[::1]/foo/bar".asURL!
-        XCTAssertTrue(url.isIPv6)
-        XCTAssertNil(url.baseDomain)
-        XCTAssertEqual(url.normalizedHost!, "[::1]")
-    }
-
     private func checkUrls(goodurls: [String], badurls: [String], checker: (InternalURL) -> Bool) {
         goodurls.forEach {
             var result = false
@@ -357,15 +249,6 @@ class NSURLExtensionsTests: XCTestCase {
         badurls.forEach { XCTAssertFalse(URL(string: $0)!.isWebPage(), $0) }
     }
 
-    func testdomainURL() {
-        let urls = [
-            ("https://www.example.com/index.html", "https://example.com/"),
-            ("https://mail.example.com/index.html", "https://mail.example.com/"),
-            ("https://mail.example.co.uk/index.html", "https://mail.example.co.uk/"),
-        ]
-        urls.forEach { XCTAssertEqual(URL(string: $0.0)!.domainURL.absoluteString, $0.1) }
-    }
-
     func testdisplayURL() {
         let goodurls = [
             ("http://localhost:\(AppInfo.webserverPort)/reader-mode/page?url=https%3A%2F%2Fen%2Em%2Ewikipedia%2Eorg%2Fwiki%2F",
@@ -386,33 +269,6 @@ class NSURLExtensionsTests: XCTestCase {
 
         goodurls.forEach { XCTAssertEqual(URL(string: $0.0)!.displayURL?.absoluteString, $0.1) }
         badurls.forEach { XCTAssertNil(URL(string: $0)!.displayURL) }
-    }
-
-    func testnormalizedHostAndPath() {
-        let goodurls = [
-            ("https://www.example.com/index.html", "example.com/index.html"),
-            ("https://mail.example.com/index.html", "mail.example.com/index.html"),
-            ("https://mail.example.co.uk/index.html", "mail.example.co.uk/index.html"),
-            ("https://m.example.co.uk/index.html", "example.co.uk/index.html")
-        ]
-        let badurls = [
-            "http:///errors/error.html",
-            "http://:\(AppInfo.webserverPort)/about/home",
-        ]
-
-        goodurls.forEach { XCTAssertEqual(URL(string: $0.0)!.normalizedHostAndPath, $0.1) }
-        badurls.forEach { XCTAssertNil(URL(string: $0)!.normalizedHostAndPath) }
-    }
-
-    func testShortDisplayString() {
-        let urls = [
-            ("https://www.example.com/index.html", "example"),
-            ("https://m.foo.com/bar/baz?noo=abc#123", "foo"),
-            ("https://user:pass@m.foo.com/bar/baz?noo=abc#123", "foo"),
-            ("https://accounts.foo.com/bar/baz?noo=abc#123", "accounts.foo"),
-            ("https://accounts.what.foo.co.za/bar/baz?noo=abc#123", "accounts.what.foo"),
-        ]
-        urls.forEach { XCTAssertEqual(URL(string: $0.0)!.shortDisplayString, $0.1) }
     }
 
     func testorigin() {
